First cut, More tweaking needed
The temperature blending on the bottom isn't quite right yet - it's way too red when it's supposed to be cold. The other tweaks are that I want to poll corelocation for the lat/long and I want to have it update on a timer. We also need to break things up to be a whole lot more readable